consider examples for parameter values
If `example` isn't set for a parameter, then go ahead
and check for `examples`. If that is present, then use the
value from the first entry as the parameter value.

Note: The value of the fist entry must have type `string`.
If no suitable value is found, then the code snippet
will not have a value.

Fix #81