create function createHarParameterObjects
This function takes an OpenAPI parameter and a sample value and
creates a HAR "parameter" object which is just a {name: "", value: ""}
object.

The function handles all of the variatons outlined in the spec
and this commit adds all the test cases that are presented
here: https://swagger.io/docs/specification/serialization

This function will be used in future commits to handle every
type of parrameter while generating snippets.