Re-order checks for candidate example param values
Per the swagger guide: https://swagger.io/docs/specification/describing-parameters/

"Using default to specify a sample value. This is not intended use of
default and can lead to unexpected behavior in some Swagger tools.
Use the example or examples keyword for this purpose instead."

Per this I think `default` should be the sample value of last resort
instead favoring `example` and `examples` first.