fix: default missing type to any type
Per the spec if no type is given then the type default to the Any Type:
https://swagger.io/docs/specification/data-models/data-types/#any

This commit removes a check that explicitly sets any undefined
type to object, instead leaving it blank. Then where type checking
is done it adds support for missing types.

One existing test was expecting SOME_OBJECT_TYPE because of a missing
type, this test was corrected to look for SOME_ANY_TYPE. An additional
test was added to explicitly state that missing types are to be
considered Any Type.

In addition a fix was made that if the value of a type field is not
a string, then the parmater value SOME_ERROR_TYPE is generated. This
allows the snippet generator to run without error, but presumably
in some workflow the openapi validator will catch the spec error.

Fixes #89